more progress

+++ JDBC_ex2/ejercicio/src/main/java/com/jdbc2/ejercicio/Models/Tractor.java
@@ -11,10 +11,10 @@ public class Tractor implements IQueriableResource {
     private int id;
     private String modelo;
     private int velocidad;
-    private int precio_venta;
+    private double precio_venta;
     private int id_construccion;
 
-    public Tractor(int id, String modelo, int velocidad, int precio_venta, int id_construccion){
+    public Tractor(int id, String modelo, int velocidad, double precio_venta, int id_construccion){
         this.id = id;
         this.modelo = modelo;
         this.velocidad = velocidad;
@@ -29,11 +29,11 @@ public class Tractor implements IQueriableResource {
     public void SetId_construccion(int id_construccion){
         this.id_construccion = id_construccion;
     }
-    public int GetPrecio_venta(){
+    public double GetPrecio_venta(){
         return this.precio_venta;
     }
     
-    public void SetPrecio_venta(int precio_venta){
+    public void SetPrecio_venta(double precio_venta){
         this.precio_venta = precio_venta;
     }
     public int GetVelocidad(){
@@ -71,7 +71,7 @@ public class Tractor implements IQueriableResource {
         statement.setInt(1, id);
         statement.setString(2, modelo);
         statement.setInt(3, velocidad);
-        statement.setInt(4, precio_venta);
+        statement.setDouble(4, precio_venta);
         statement.setInt(5, id_construccion);
         return statement;
     }
